+++ Changelog.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0]
+### Added
+-Javafx functionality
+
+-Student, Course classes to store data
+
+-IOManager Class to read and write data
+
+
+###Changed
+
+-Created a functional GUI that can load data from a file and display
+it as specified in the HW2 requirements.
+
+-Updated all documentation
+
 ## [1.0.0] - 2021-10-01
 ### Added
 -Readme
@@ -14,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -HelloWorld
 
+-App
 ### Changed
 N/A